% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cashflows}
\alias{cashflows}
\title{Example of cashflows}
\format{An example dataset with 5 rows and 2 variables:
\describe{
 \item{cfdate}{date, in the form of yyyy-mm-dd, as a lubridate package object}
 \item{cfamount}{cash flow, outgoing cash flows as negative number, incoming cashflows as positive number}
 }


 library(lubridate)
 library(plyr)
 cfdate <- ymd(c("2019-08-03", "2023-02-03", "2022-08-03", "2024-04-03", "2025-03-03"))
 cfamount <- c(-51000000, 30000000, 90000000, -22000000, 10000000)
 cashflows <- data.frame(cfdate, cfamount)}
\usage{
cashflows
}
\description{
This is an example of how to create the cashflows data frame which is used as an input parameter for \code{xnpv} and \code{xirr}
}
\keyword{datasets}
