% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xirr.R
\name{xirr}
\alias{xirr}
\title{xirr(cashflows, guess, tol, m)}
\usage{
xirr(cashflows, guess, tol, m)
}
\arguments{
\item{guess}{(optional, default = 0.1): a guess at the solution to be used as a starting point for the numerical solution.}

\item{tol}{(optional, default = 0.000000001): the error tolerance.}

\item{m}{(optional, default = 500): the maxmimum number (integer) of iterations used for the secant method.}

\item{cashflows:}{a tibble in which each element is of the form (date, amount), where date is an r lubridate object and amount is an integer or floating point number. Cash outflows (investments) are represented with negative amounts, and cash inflows (returns) are positive amounts.}
}
\value{
Returns the IRR as a single value (floating point number)
}
\description{
Calculates the Internal Rate of Return of a series of cashflows at irregular intervals.
}
\note{
The Internal Rate of Return (IRR) is the discount rate at which the Net Present Value (NPV) of a series of cash flows is equal to zero. The NPV of the series of cash flows is determined using \code{xnpv} function of this package. The discount rate at which NPV equals zero is found using the secant method of the cmna package. This \code{xirr} function is equivalent to the Microsoft Excel function of the same name.
}
